+++ PoshBot/Classes/Scheduler.ps1
@@ -34,10 +34,16 @@ class Scheduler : BaseLogger {
                 } else {
                     if (-not [string]::IsNullOrEmpty($sched.StartAfter)) {
                         $newSchedule = [ScheduledMessage]::new($sched.TimeInterval, $sched.TimeValue, $msg, $sched.Enabled, $sched.StartAfter.ToUniversalTime())
+
+                        if ($newSchedule.StartAfter -lt (Get-Date).ToUniversalTime()) {
+                            #Prevent reruns of commands initially scheduled at least one interval ago
+                            $newSchedule.RecalculateStartAfter()
+                        }
                     } else {
                         $newSchedule = [ScheduledMessage]::new($sched.TimeInterval, $sched.TimeValue, $msg, $sched.Enabled, (Get-Date).ToUniversalTime())
                     }
                 }
+
                 $newSchedule.Id = $sched.Id
                 $this.ScheduleMessage($newSchedule, $false)
             }
@@ -96,7 +102,6 @@ class Scheduler : BaseLogger {
 
                 # Check if one time command
                 if ($_.Value.Once) {
-                    $_.Value.StopTimer()
                     $remove += $_.Value.Id
                 } else {
                     $_.Value.RecalculateStartAfter()