+++ PoshBot/Public/Remove-PoshbotStatefulData.ps1
@@ -40,11 +40,11 @@ function Remove-PoshBotStatefulData {
     )
     process {
         if($Scope -eq 'Module') {
-            $FileName = "$($PoshBotContext.Plugin).state"
+            $FileName = "$($global:PoshBotContext.Plugin).state"
         } else {
             $FileName = "PoshbotGlobal.state"
         }
-        $Path = Join-Path $PoshBotContext.ConfigurationDirectory $FileName
+        $Path = Join-Path $global:PoshBotContext.ConfigurationDirectory $FileName
 
 
         if(-not (Test-Path $Path)) {