+++ PoshBot/Public/Start-PoshBot.ps1
@@ -92,10 +92,11 @@ function Start-PoshBot {
                 $sb = {
                     param(
                         [parameter(Mandatory)]
-                        [hashtable]$Configuration
+                        [hashtable]$Configuration,
+                        [string]$PoshBotManifestPath
                     )
 
-                    Import-Module PoshBot -ErrorAction Stop
+                    Import-Module $PoshBotManifestPath -ErrorAction Stop
 
                     try {
                         $tempConfig = New-PoshBotConfiguration
@@ -103,12 +104,12 @@ function Start-PoshBot {
 
                         while ($true) {
                             try {
-                                if ($realConfig.BackendConfiguration.Name -eq 'Slack') {
+                                if ($realConfig.BackendConfiguration.Name -in @('Slack', 'SlackBackend')) {
                                     $backend = New-PoshBotSlackBackend -Configuration $realConfig.BackendConfiguration
-                                } elseIf ($realConfig.BackendConfiguration.Name -eq 'Teams') {
+                                } elseIf ($realConfig.BackendConfiguration.Name -in @('Teams', 'TeamsBackend')) {
                                     $backend = New-PoshBotTeamsBackend -Configuration $realConfig.BackendConfiguration
                                 } else {
-                                    Write-Error "Unable to determine backend type. Name property in BackendConfiguration should have a value of 'Slack' or 'Teams'"
+                                    Write-Error "Unable to determine backend type. Name property in BackendConfiguration should have a value of 'Slack', 'SlackBackend', 'Teams', or 'TeamsBackend'"
                                     break
                                 }
 
@@ -127,8 +128,9 @@ function Start-PoshBot {
 
                 $instanceId = (New-Guid).ToString().Replace('-', '')
                 $jobName = "PoshBot_$instanceId"
+                $poshBotManifestPath = (Join-Path -Path $PSScriptRoot -ChildPath "PoshBot.psd1")
 
-                $job = Start-Job -ScriptBlock $sb -Name $jobName -ArgumentList $Configuration.ToHash()
+                $job = Start-Job -ScriptBlock $sb -Name $jobName -ArgumentList $Configuration.ToHash(),$poshBotManifestPath
 
                 # Track the bot instance
                 $botTracker = @{