+++ Tests/Unit/Classes/ScheduledMessages.tests.ps1
@@ -100,11 +100,11 @@ InModuleScope PoshBot {
             $ScheduledMessage = [ScheduledMessage]::new($Interval, $TimeValue, $Message, (Get-Date).AddHours(5))
 
             It 'Should return true when past the StartAfter DateTime' {
-                $ElapsedMessage.HasElapsed() | Should BeTrue
+                $ElapsedMessage.HasElapsed() | Should be $true
             }
 
             It 'Should return false when before the StartAfter DateTime' {
-                $ScheduledMessage.HasElapsed() | Should Not BeTrue
+                $ScheduledMessage.HasElapsed() | Should Not be $true
             }
 
             It 'TimesExecuted should be 1 after executing' {
@@ -118,12 +118,12 @@ InModuleScope PoshBot {
 
             It 'Disables the instance when called' {
                 $ScheduledMessage.Disable()
-                $SceduleMessage.Enabled | Should Not BeTrue
+                $SceduleMessage.Enabled | Should Not be $true
             }
 
             It 'Does not throw an error when being called on an already-disabled instance' {
                 $ScheduledMessage.Disable()
-                $ScheduledMessage.Enabled | Should Not BeTrue
+                $ScheduledMessage.Enabled | Should Not be $true
             }
         }
 
@@ -133,23 +133,42 @@ InModuleScope PoshBot {
             It 'Enables the instance when called' {
                 $ScheduledMessage.Disable()
                 $ScheduledMessage.Enable()
-                $ScheduledMessage.Enabled | Should BeTrue
+                $ScheduledMessage.Enabled | Should be $true
             }
 
             It 'Does not throw an error when being called on an already-enabled instance' {
                 $ScheduledMessage.Enable()
-                $ScheduledMessage.Enabled | Should BeTrue
+                $ScheduledMessage.Enabled | Should be $true
             }
         }
 
         Context "Method: RecalculateStartAfter()" {
-            $ScheduledMessage = [ScheduledMessage]::new($Interval, $TimeValue, $Message, (Get-Date))
-
             It 'Should increase the StartAfter property by IntervalMS' {
+                $ScheduledMessage = [ScheduledMessage]::new($Interval, $TimeValue, $Message, (Get-Date))
+
                 $StartingValue = $ScheduledMessage.StartAfter
                 $ScheduledMessage.RecalculateStartAfter()
 
                 (New-TimeSpan $StartingValue $ScheduledMessage.StartAfter).TotalDays | Should Be 1
+            }
+
+            It 'Should not reschedule a run before the current time' {
+                $currentDate = (Get-Date).ToUniversalTime();
+                $startAfter = $currentDate.AddDays(-5);
+                $ScheduledMessage = [ScheduledMessage]::new($Interval, $TimeValue, $Message, $startAfter)
+
+                $ScheduledMessage.RecalculateStartAfter();
+
+                $ScheduledMessage.StartAfter | Should Not BeLessThan $currentDate
+            }
+
+            It 'Should only move StartAfter value forward' {
+                $StartingValue = (Get-Date).AddDays($daysDifference).ToUniversalTime()
+                $ScheduledMessage = [ScheduledMessage]::new($Interval, $TimeValue, $Message, $StartingValue)
+
+                $ScheduledMessage.RecalculateStartAfter()
+
+                $ScheduledMessage.StartAfter | Should Be $StartingValue.AddDays(1)
             }
         }