+++ Tests/Unit/Public/New-PoshBotFileUpload.tests.ps1
@@ -2,8 +2,7 @@
 describe 'New-PoshBotFileUpload' {
 
     BeforeAll {
-        $readme = $env:bh
-        $readme = Join-Path -Path $env:BHProjectPath -ChildPath 'readme.md'
+        $readme = Join-Path -Path $env:BHProjectPath -ChildPath 'README.md'
     }
 
     it 'Returns a [PoshBot.File.Upload] object' {
@@ -17,8 +16,8 @@ describe 'New-PoshBotFileUpload' {
     }
 
     it 'Has a valid [Title] field' {
-        $resp = New-PoshBotFileUpload -Path $readme -DM -Title 'readme.md'
-        $resp.Title | should be 'readme.md'
+        $resp = New-PoshBotFileUpload -Path $readme -DM -Title 'README.md'
+        $resp.Title | should be 'README.md'
     }
 
     it 'Validates file exists' {
@@ -26,5 +25,27 @@ describe 'New-PoshBotFileUpload' {
         $badFile = Join-Path -Path $env:BHProjectPath -ChildPath "$($guid).txt"
         { New-PoshBotFileUpload -Path $badFile } | should throw
     }
+
+    it 'Sends file content and not path' {
+        $resp = New-PoshBotFileUpload -Content 'foo'
+        $resp.Path | should be $null
+    }
+
+    it 'Has a valid [FileName] field when sending file content' {
+        $resp = New-PoshBotFileUpload -Content 'foo' -FileName 'foo.txt'
+        $resp.FileName | should be 'foo.txt'
+    }
+
+    it 'Has a valid [FileType] field when sending file content' {
+        $resp = New-PoshBotFileUpload -Content 'foo' -FileType 'powershell'
+        $resp.FileType | should be 'powershell'
+    }
+
+    it 'Sends file path and not content' {
+        $resp = New-PoshBotFileUpload -Path $readme
+        $resp.Content | should be $null
+    }
+
+
 }