+++ Tests/Unit/Public/Remove-PoshBotStatefulData.tests.ps1
@@ -5,14 +5,20 @@ InModuleScope PoshBot {
             $PSDefaultParameterValues = @{
                 'Remove-PoshBotStatefulData:Verbose' = $false
             }
+
+            # Define internal variables
+            $global:PoshBotContext = [pscustomobject]@{
+                Plugin                 = 'TestPlugin'
+                ConfigurationDirectory = (Join-Path $env:BHProjectPath Tests)
+            }
         }
-        # Define internal variables
-        $poshbotcontext = [pscustomobject]@{
-            Plugin = 'TestPlugin'
-            ConfigurationDirectory = (Join-Path $env:BHProjectPath Tests)
+
+        AfterAll {
+            Remove-Variable -Name PoshBotContext -Scope Global -Force
         }
-        $globalfile = Join-Path $poshbotcontext.ConfigurationDirectory "PoshbotGlobal.state"
-        $modulefile = Join-Path $poshbotcontext.ConfigurationDirectory "$($poshbotcontext.Plugin).state"
+
+        $globalfile = Join-Path $global:PoshBotContext.ConfigurationDirectory 'PoshbotGlobal.state'
+        $modulefile = Join-Path $global:PoshBotContext.ConfigurationDirectory "$($poshbotcontext.Plugin).state"
         [pscustomobject]@{
             a = 'g'
             b = $true