+++ build.ps1
@@ -1,91 +1,39 @@
 
 [cmdletbinding(DefaultParameterSetName = 'task')]
 param(
+    # Build task(s) to execute
     [parameter(ParameterSetName = 'task', Position = 0)]
     [string[]]$Task = 'default',
 
-    [parameter(ParameterSetName = 'help')]
-    [switch]$Help,
+    # Bootstrap dependencies
+    [switch]$Bootstrap,
 
-    [switch]$UpdateModules
+    # List available build tasks
+    [parameter(ParameterSetName = 'Help')]
+    [switch]$Help
 )
 
-function Resolve-Module {
-    [Cmdletbinding()]
-    param (
-        [Parameter(Mandatory, ValueFromPipeline)]
-        [string[]]$Name,
+$ErrorActionPreference = 'Stop'
 
-        [switch]$UpdateModules
-    )
-
-    begin {
-        Get-PackageProvider -Name Nuget -ForceBootstrap -Verbose:$false | Out-Null
-        Set-PSRepository -Name PSGallery -InstallationPolicy Trusted -Verbose:$false
-
-        $PSDefaultParameterValues = @{
-            '*-Module:Verbose' = $false
-            'Install-Module:ErrorAction' = 'Stop'
-            'Install-Module:Force' = $true
-            'Install-Module:Scope' = 'CurrentUser'
-            'Install-Module:Verbose' = $false
-            'Install-Module:AllowClobber' = $true
-            'Import-Module:ErrorAction' = 'Stop'
-            'Import-Module:Verbose' = $false
-            'Import-Module:Force' = $true
-        }
+# Bootstrap dependencies
+if ($Bootstrap.IsPresent) {
+    Get-PackageProvider -Name Nuget -ForceBootstrap > $null
+    Set-PSRepository -Name PSGallery -InstallationPolicy Trusted
+    if (-not (Get-Module -Name PSDepend -ListAvailable)) {
+        Install-Module -Name PSDepend -Repository PSGallery -Scope CurrentUser -Force
     }
-
-    process {
-        foreach ($moduleName in $Name) {
-            $versionToImport = ''
-
-            Write-Verbose -Message "Resolving Module [$($moduleName)]"
-            if ($Module = Get-Module -Name $moduleName -ListAvailable -Verbose:$false) {
-                # Determine latest version on PSGallery and warn us if we're out of date
-                $latestLocalVersion = ($Module | Measure-Object -Property Version -Maximum).Maximum
-                $latestGalleryVersion = (Find-Module -Name $moduleName -Repository PSGallery |
-                    Measure-Object -Property Version -Maximum).Maximum
-
-                # Out we out of date?
-                if ($latestLocalVersion -lt $latestGalleryVersion) {
-                    if ($UpdateModules) {
-                        Write-Verbose -Message "$($moduleName) installed version [$($latestLocalVersion.ToString())] is outdated. Installing gallery version [$($latestGalleryVersion.ToString())]"
-                        if ($UpdateModules) {
-                            Install-Module -Name $moduleName -RequiredVersion $latestGalleryVersion
-                            $versionToImport = $latestGalleryVersion
-                        }
-                    } else {
-                        Write-Warning "$($moduleName) is out of date. Latest version on PSGallery is [$latestGalleryVersion]. To update, use the -UpdateModules switch."
-                    }
-                } else {
-                    $versionToImport = $latestLocalVersion
-                }
-            } else {
-                Write-Verbose -Message "[$($moduleName)] missing. Installing..."
-                Install-Module -Name $moduleName -Repository PSGallery
-                $versionToImport = (Get-Module -Name $moduleName -ListAvailable | Measure-Object -Property Version -Maximum).Maximum
-            }
-
-            Write-Verbose -Message "$($moduleName) installed. Importing..."
-            if (-not [string]::IsNullOrEmpty($versionToImport)) {
-                Import-module -Name $moduleName -RequiredVersion $versionToImport
-            } else {
-                Import-module -Name $moduleName
-            }
-        }
-    }
+    Import-Module -Name PSDepend -Verbose:$false
+    Invoke-PSDepend -Path './requirements.psd1' -Install -Import -Force -WarningAction SilentlyContinue
 }
 
-'BuildHelpers', 'psake' | Resolve-Module -UpdateModules:($PSBoundParameters.ContainsKey('UpdateModules'))
-
-if ($PSBoundParameters.ContainsKey('help')) {
-    Get-PSakeScriptTasks -buildFile "$PSScriptRoot\psake.ps1" |
-        Sort-Object -Property Name |
+# Execute psake task(s)
+$psakeFile = './psake.ps1'
+if ($PSCmdlet.ParameterSetName -eq 'Help') {
+    Get-PSakeScriptTasks -buildFile $psakeFile  |
         Format-Table -Property Name, Description, Alias, DependsOn
 } else {
     Set-BuildEnvironment -Force
 
-    Invoke-psake -buildFile "$PSScriptRoot\psake.ps1" -taskList $Task -nologo -Verbose:($VerbosePreference -eq 'Continue')
+    Invoke-psake -buildFile $psakeFile -taskList $Task -nologo
     exit ( [int]( -not $psake.build_success ) )
 }